Update the structure of the project to enable importing classes from other package.
