+++ S2SRL/SymbolicExecutor/change.py
@@ -8,7 +8,7 @@
 '''
 
 import json
-from SymbolicExecutor.symbolics import Symbolics
+from S2SRL.SymbolicExecutor import symbolics
 
 
 def transMask2Action(state):
@@ -54,7 +54,7 @@ def transMask2Action(state):
                 #     symbolic_seq[-1] = {"A3":["","",""]} if not symbolic_seq[-1].has_key("A3") else symbolic_seq[-1]### A3
                 # if state.startswith("QuantitativeReasoning(Count)(All)") or state.startswith("ComparativeReasoning(Count)(All)"):
                 #     symbolic_seq[-1] = {"A11": ["", "", ""]} if not symbolic_seq[-1].has_key("A11") else symbolic_seq[-1]
-                symbolic_exe = Symbolics(symbolic_seq)
+                symbolic_exe = symbolics.Symbolics(symbolic_seq)
                 answer = symbolic_exe.executor()
 
                 if state.startswith("QuantitativeReasoning(Count)(All)") or state.startswith("ComparativeReasoning(Count)(All)"):
