+++ S2SRL/SymbolicExecutor/test.py
@@ -9,7 +9,7 @@ import copy
 import time
 
 from Preprocess.load_qadata import load_qadata, getQA_by_state
-from symbolics import Symbolics
+from .symbolics import Symbolics
 import logging
 logging.basicConfig(level=logging.INFO,#控制台打印的日志级别
                     filename='/data/zjy/count_orig.log',
