+++ S2SRL/__init__.py
@@ -1,0 +1,6 @@
+# -*- coding: utf-8 -*-
+# @Time    : 2019/8/31 05:18 AM
+# To set the root path fot import class from other packages.
+import os
+import sys
+sys.path.append(os.path.realpath(os.getcwd()))
